Replaced all uses of ViewModelBase with ObservableBase. Removed ViewModelBase. Moved MainWindowViewModel into a separate namespace and split its file.
