+++ AnnoMapEditor/UI/MainWindow.xaml.cs
@@ -7,6 +7,7 @@ using System.Diagnostics;
 using System.Windows;
 using System.Windows.Controls;
 using AnnoMapEditor.Utilities;
+using AnnoMapEditor.UI.Models.MainWindow;
 
 namespace AnnoMapEditor.UI
 {
