+++ AnnoMapEditor/UI/Models/MainWindow/ExportStatus.cs
@@ -1,0 +1,8 @@
+﻿namespace AnnoMapEditor.UI.Models
+{
+    public class ExportStatus
+    {
+        public bool CanExportAsMod { get; set; }
+        public string ExportAsModText { get; set; } = "";
+    }
+}
