+++ AnnoMapEditor/UI/Models/MainWindow/MapGroup.cs
@@ -1,0 +1,24 @@
+﻿using System.Collections.Generic;
+using System.Globalization;
+using System.Linq;
+using System.Text.RegularExpressions;
+
+namespace AnnoMapEditor.UI.Models
+{
+    public class MapGroup
+    {
+        public string Name;
+        public List<MapInfo> Maps;
+
+        public MapGroup(string name, IEnumerable<string> mapFiles, Regex regex)
+        {
+            Name = name;
+            Maps = mapFiles.Select(x => new MapInfo()
+            {
+                Name = CultureInfo.CurrentCulture.TextInfo.ToTitleCase(
+                    string.Join(' ', regex.Match(x).Groups.Values.Skip(1).Select(y => y.Value)).Replace("_", " ")),
+                FileName = x
+            }).ToList();
+        }
+    }
+}
