+++ AnnoMapEditor/UI/Models/MainWindow/MapInfo.cs
@@ -1,0 +1,8 @@
+﻿namespace AnnoMapEditor.UI.Models
+{
+    public class MapInfo
+    {
+        public string? Name;
+        public string? FileName;
+    }
+}
